[FLINK-37357][table-planner]Migrating SqlNode conversion logic from SqlNodeToOperationConversion to SqlNodeConverters
