+++ flink-table/flink-table-planner/src/main/java/org/apache/flink/table/planner/operations/converters/SqlAlterFunctionConverter.java
@@ -1,0 +1,50 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.table.planner.operations.converters;
+
+import org.apache.flink.sql.parser.ddl.SqlAlterFunction;
+import org.apache.flink.table.api.ValidationException;
+import org.apache.flink.table.catalog.CatalogFunction;
+import org.apache.flink.table.catalog.CatalogFunctionImpl;
+import org.apache.flink.table.catalog.FunctionLanguage;
+import org.apache.flink.table.catalog.ObjectIdentifier;
+import org.apache.flink.table.catalog.UnresolvedIdentifier;
+import org.apache.flink.table.operations.Operation;
+import org.apache.flink.table.operations.ddl.AlterCatalogFunctionOperation;
+
+public class SqlAlterFunctionConverter implements SqlNodeConverter<SqlAlterFunction> {
+    @Override
+    public Operation convertSqlNode(SqlAlterFunction node, ConvertContext context) {
+        if (node.isSystemFunction()) {
+            throw new ValidationException("Alter temporary system function is not supported");
+        }
+
+        FunctionLanguage language = SqlNodeConvertUtils.parseLanguage(node.getFunctionLanguage());
+        CatalogFunction catalogFunction =
+                new CatalogFunctionImpl(
+                        node.getFunctionClassName().getValueAs(String.class), language);
+
+        UnresolvedIdentifier unresolvedIdentifier =
+                UnresolvedIdentifier.of(node.getFunctionIdentifier());
+        ObjectIdentifier identifier =
+                context.getCatalogManager().qualifyIdentifier(unresolvedIdentifier);
+        return new AlterCatalogFunctionOperation(
+                identifier, catalogFunction, node.isIfExists(), node.isTemporary());
+    }
+}
