+++ flink-table/flink-table-planner/src/main/java/org/apache/flink/table/planner/operations/converters/SqlCompileAndExecutePlanConverter.java
@@ -1,0 +1,37 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.table.planner.operations.converters;
+
+import org.apache.flink.sql.parser.dml.SqlCompileAndExecutePlan;
+import org.apache.flink.table.operations.CompileAndExecutePlanOperation;
+import org.apache.flink.table.operations.Operation;
+import org.apache.flink.table.planner.operations.SqlNodeToOperationConversion;
+
+public class SqlCompileAndExecutePlanConverter
+        implements SqlNodeConverter<SqlCompileAndExecutePlan> {
+    @Override
+    public Operation convertSqlNode(SqlCompileAndExecutePlan node, ConvertContext context) {
+        return new CompileAndExecutePlanOperation(
+                node.getPlanFile(),
+                SqlNodeToOperationConversion.convertValidatedSqlNodeOrFail(
+                        context.getFlinkPlanner(),
+                        context.getCatalogManager(),
+                        node.getOperandList().get(0)));
+    }
+}
