+++ flink-table/flink-table-planner/src/main/java/org/apache/flink/table/planner/operations/converters/SqlLoadModuleConverter.java
@@ -1,0 +1,43 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.table.planner.operations.converters;
+
+import org.apache.flink.sql.parser.ddl.SqlTableOption;
+import org.apache.flink.sql.parser.dql.SqlLoadModule;
+import org.apache.flink.table.operations.LoadModuleOperation;
+import org.apache.flink.table.operations.Operation;
+
+import org.apache.calcite.sql.SqlNode;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class SqlLoadModuleConverter implements SqlNodeConverter<SqlLoadModule> {
+
+    @Override
+    public Operation convertSqlNode(SqlLoadModule node, ConvertContext context) {
+        String moduleName = node.moduleName();
+        Map<String, String> properties = new HashMap<>();
+        for (SqlNode sqlNode : node.getPropertyList().getList()) {
+            SqlTableOption option = (SqlTableOption) sqlNode;
+            properties.put(option.getKeyString(), option.getValueString());
+        }
+        return new LoadModuleOperation(moduleName, properties);
+    }
+}
